Request parsing dispatcher moved to its own crate
Reading Kafka's impl > Reading Kafka's doc.
ApiKey-based request parsing dispatcher
