+++ src/main.rs
@@ -5,6 +5,7 @@
 extern crate nom;
 
 use parser::primitive::*;
+use parser::request::*;
 use parser::metadata::*;
 
 mod parser;
